略

+++ Config.cc
@@ -10,12 +10,12 @@ const void *DECLINED_ADDR_token = "RC_DECLINED";
 static_assert(LARGE_CLIENT_HEADER_BUFFER_SIZE > CLIENT_HEADER_BUFFER_SIZE, 
 "LARGE_CLIENT_HEADER_BUFFER_SIZE 应大于 CLIENT_HEADER_BUFFER_SIZE");
 
-unsigned hash(size_t key, u_char c)
+size_t hash(size_t key, u_char c)
 {
     return key * 31 + c;
 }
 
-unsigned hash_string(const char *c) {
+size_t hash_string(const char *c) {
     const char * pos = c;
 
     unsigned ans = 0;
